+++ src/pages/UsersPage/UserSearchPage.jsx
@@ -1,22 +1,30 @@
 import React from 'react';
 import styled from 'styled-components';
-import UserSearchSideSection from '../../components/UsersPage/UserSearchSideSection';
+import SideSection from '../../components/common/SideSection';
+import SearchBar from '../../components/UsersPage/SearchBar';
+import UserInfo from '../../components/UsersPage/UserInfo';
+
+
 
 const UserSearchPage = () => {
   return (
-    <UserSearchContainer>
-      <UserSearchSideSection />
-    </UserSearchContainer>
+    <SideSection>
+        <ContentBox>
+          <SearchBar placeholder="유저 핸들을 검색" />
+        </ContentBox>
+        <ContentBox>
+          <UserInfo />
+        </ContentBox>
+    </SideSection>
   );
 };
 
 export default UserSearchPage;
 
-const UserSearchContainer = styled.div``;
-
-const Content = styled.div`
-  margin: 33px;
+const ContentBox = styled.div`
+  padding: 33px;
   padding-top: 40px;
-  display: flex;
-  flex-direction: column;
+  border-bottom: 1px solid var(--gray, #bcbcbc);
 `;
+
+
